# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "US College Application Consultant API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
...
      "PromptResponse": {
        "type": "object",
        "properties": {
          "prompt": {
            "type": "string"
          },
          "hasDeepSeekApiKey": {
            "type": "boolean"
          }
        },
        "required": [
          "prompt",
          "hasDeepSeekApiKey"
        ]
      },
      "RagStreamRequest": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1200
          },
          "historySummary": {
            "type": "string",
            "maxLength": 1800,
            "default": ""
          },
          "assistantProfile": {
            "type": "string",
            "enum": [
              "",
              "major-match",
              "inspiration"
            ],
            "default": ""
          },
          "usePersonalContext": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "question"
        ]
      },
...
      "UnifiedError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "requestId": {
            "type": "string",
            "maxLength": 128
          },
          "retryable": {
            "type": "boolean",
            "default": false
          },
          "details": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "error",
          "code"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/prompt": {
      "get": {
        "responses": {
          "200": {
            "description": "Fixed planning prompt and model configuration status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unified API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedError"
                }
              }
            }
          },
          "500": {
            "description": "Unified API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/deepseek-rag/stream": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RagStreamRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "RAG progress and final result as Server-Sent Events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Unified API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedError"
                }
              }
            }
          },
          "401": {
            "description": "Unified API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedError"
                }
              }
            }
          },
          "403": {
            "description": "Unified API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedError"
                }
              }
            }
          },
          "429": {
            "description": "Unified API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedError"
                }
              }
            }
          },
          "500": {
            "description": "Unified API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedError"
                }
              }
            }
          }
        }
      }
    }
  },
...
}